hago strmapi, striteri, putnbr, putendl, y añado modificaciones para que paco de OK

+++ ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: kcanales <kcanales@student.42madrid.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/28 10:54:12 by kcanales          #+#    #+#             */
-/*   Updated: 2025/05/16 16:29:25 by kcanales         ###   ########.fr       */
+/*   Updated: 2025/05/23 15:48:29 by kcanales         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *s)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i] != '\0')
